Changed the attributes of the Point class from doubles to ints

+++ Point.java
@@ -1,16 +1,16 @@
 public class Point{
-   private double x;
-   private double y;
-   public Point(double x, double y){
+   private int x;
+   private int y;
+   public Point(int x, int y){
       this.x = x;
       this.y = y;
    }
 
-   public double xCoord(){
+   public int xCoord(){
       return this.x;
    }
 
-   public double yCoord(){
+   public int yCoord(){
       return this.y;
    }
 }
